CommunicationManager: added single lock during the tag-based fence
The zero-cost fences make sense to not use locks, but this "heavyweight" variation does not.
What was currently happening (in MPI) was the loop over slots doing a lock-unlock in each iteration.
We eliminate all this with a global single lock.

When the simplification that removes all redundant Win_fence's from the fenceImpl() is merged, we expect a conflict
here so this commit should be fixed accordingly