CommunicationManager: [MPI] Added deregisterGlobalMemorySlot specific implementation
We discovered that with MPI, if we don't keep track of all slots ever exchanged, it is
possible that we lose reference to the MPI_Window associated with a slot (if the slot is de-registered etc.)
That means that when the -collective- fence attempts to do collective slot destruction the system will deadlock,
if even just one instace does not have access to the Windows the slot was created with.

That will not be a problem in other backends (e.g., in LPF as the slot is just an ID we can simply communicate
it during the fence and the collective deregister can happen without issues). However, in MPI we need to still keep
track of deregistered slots, so we create an additional, shadow map of _globalMemorySlotTagKeyMap where we move
the slots that 'deregister' is called upon. This way we achieve the desired effect (getGlobalSlot() does not work on them)
but we can participate in a collective slot destruction...