examples: updated all relevant examples with the destroyGlobalMemorySlot and subsequent fence()
Now in multi-instance scenarios different instances can call one-sided destroy on Global Slots,
there is no need to have them all do it collectively and in the same order. A collective fence is enough
to accomplish this. So now different instances may free different (or sometimes overlapping) slots
in order to demonstrate this.